+++ cmd/pull.go
@@ -1,0 +1,77 @@
+package cmd
+
+import (
+	"fmt"
+	"github.com/jmorganca/ollama/api"
+	"github.com/jmorganca/ollama/progress"
+	"github.com/spf13/cobra"
+	"os"
+)
+
+var pullCmd = &cobra.Command{
+	Use:     "pull MODEL",
+	Short:   "Pull a model from a registry",
+	Args:    cobra.ExactArgs(1),
+	PreRunE: checkServerHeartbeat,
+	RunE:    PullHandler,
+}
+
+func init() {
+	pullCmd.Flags().Bool("insecure", false, "Use an insecure registry")
+	appendHostEnvDocs(pullCmd)
+	rootCmd.AddCommand(pullCmd)
+}
+
+func PullHandler(cmd *cobra.Command, args []string) error {
+	insecure, err := cmd.Flags().GetBool("insecure")
+	if err != nil {
+		return err
+	}
+
+	client, err := api.ClientFromEnvironment()
+	if err != nil {
+		return err
+	}
+
+	p := progress.NewProgress(os.Stderr)
+	defer p.Stop()
+
+	bars := make(map[string]*progress.Bar)
+
+	var status string
+	var spinner *progress.Spinner
+
+	fn := func(resp api.ProgressResponse) error {
+		if resp.Digest != "" {
+			if spinner != nil {
+				spinner.Stop()
+			}
+
+			bar, ok := bars[resp.Digest]
+			if !ok {
+				bar = progress.NewBar(fmt.Sprintf("pulling %s...", resp.Digest[7:19]), resp.Total, resp.Completed)
+				bars[resp.Digest] = bar
+				p.Add(resp.Digest, bar)
+			}
+
+			bar.Set(resp.Completed)
+		} else if status != resp.Status {
+			if spinner != nil {
+				spinner.Stop()
+			}
+
+			status = resp.Status
+			spinner = progress.NewSpinner(status)
+			p.Add(status, spinner)
+		}
+
+		return nil
+	}
+
+	request := api.PullRequest{Name: args[0], Insecure: insecure}
+	if err := client.Pull(cmd.Context(), &request, fn); err != nil {
+		return err
+	}
+
+	return nil
+}
